Class renaming and package refactoring
