+++ src/main/scala/org/gammf/collabora/EntryPoint.scala
@@ -5,6 +5,7 @@ import com.newmotion.akka.rabbitmq.{ConnectionActor, ConnectionFactory}
 import org.gammf.collabora.authentication.AuthenticationServer
 import org.gammf.collabora.authentication.actors.AuthenticationActor
 import org.gammf.collabora.communication.actors._
+import org.gammf.collabora.communication.actors.rabbitmq._
 import org.gammf.collabora.database.actors.ConnectionManagerActor
 import org.gammf.collabora.database.actors.master._
 import org.gammf.collabora.database.actors.worker._
@@ -25,20 +26,20 @@ object EntryPoint extends App {
   rootYellowPages ! RegistrationRequestMessage(rabbitConnection, "RabbitConnection", Topic() :+ Communication :+ RabbitMQ, ConnectionHandler)
 
   val channelCreator = system.actorOf(Props(
-    new ChannelCreatorActor(rootYellowPages, "RabbitChannelCreator", Topic() :+ Communication :+ RabbitMQ, ChannelCreating)), "channelCreator")
+    new RabbitMQChannelCreatorActor(rootYellowPages, "RabbitChannelCreator", Topic() :+ Communication :+ RabbitMQ, ChannelCreating)), "channelCreator")
   val namingActor = system.actorOf(Props(
     new RabbitMQNamingActor(rootYellowPages, "NamingActor", Topic() :+ Communication :+ RabbitMQ, Naming)), "naming")
   val publisherActor = system.actorOf(Props(
-    new PublisherActor(rootYellowPages, "PublisherActor", Topic() :+ Communication :+ RabbitMQ, Publishing)), "publisher")
+    new RabbitMQPublisherActor(rootYellowPages, "PublisherActor", Topic() :+ Communication :+ RabbitMQ, Publishing)), "publisher")
   val subscriber = system.actorOf(Props(
-    new SubscriberActor(rootYellowPages, "SubscriberActor", Topic() :+ Communication :+ RabbitMQ, Subscribing)), "subscriber")
+    new RabbitMQSubscriberActor(rootYellowPages, "SubscriberActor", Topic() :+ Communication :+ RabbitMQ, Subscribing)), "subscriber")
 
   val updatesReceiver = system.actorOf(Props
-  (new UpdatesReceiverActor(rootYellowPages, "UpdatesReceiver", Topic() :+ Communication :+ Updates :+ RabbitMQ , Master)), "updates-receiver")
+  (new RabbitMQUpdatesReceiverActor(rootYellowPages, "UpdatesReceiver", Topic() :+ Communication :+ Updates :+ RabbitMQ , Master)), "updates-receiver")
   val notificationActor = system.actorOf(Props(
-    new NotificationsSenderActor(rootYellowPages, "NotificationActor", Topic() :+ Communication :+ Notifications :+ RabbitMQ, Master)), "notifications-actor")
+    new RabbitMQNotificationsSenderActor(rootYellowPages, "NotificationActor", Topic() :+ Communication :+ Notifications :+ RabbitMQ, Master)), "notifications-actor")
   val collaborationActor = system.actorOf(Props(
-    new CollaborationMembersActor(rootYellowPages, "CollaborationActor", Topic() :+ Communication :+ Collaborations  :+ RabbitMQ, Master)), "collaborations-acotr")
+    new RabbitMQCollaborationMembersActor(rootYellowPages, "CollaborationActor", Topic() :+ Communication :+ Collaborations  :+ RabbitMQ, Master)), "collaborations-acotr")
 
   val notificationDispatcherActor = system.actorOf(Props(
     new NotificationsDispatcherActor(rootYellowPages, "NotificationDispatcher", Topic() :+ Communication :+ Notifications, Bridging)
