+++ src/main/scala/org/gammf/collabora/authentication/package.scala
@@ -8,7 +8,7 @@ package object authentication {
   /**
     * The IP where the server is listening, presumably the IP of the machine that contain the system
     */
-  val SERVER_IP_ADDRESS: String = "localhost"
+  val SERVER_IP_ADDRESS: String = "192.168.1.232"
 
   /**
     * The port where the server listen.
