+++ src/main/scala/org/gammf/collabora/communication/actors/rabbitmq/RabbitMQPublisherActor.scala
@@ -1,4 +1,4 @@
-package org.gammf.collabora.communication.actors
+package org.gammf.collabora.communication.actors.rabbitmq
 
 import akka.actor.ActorRef
 import org.gammf.collabora.communication.messages.PublishMessage
@@ -10,8 +10,8 @@ import play.api.libs.json.Json
 /**
   * This is an actor that publish a message in a certain exchange through a rabbitMQ channel.
   */
-class PublisherActor(override val yellowPages: ActorRef, override val name: String,
-                     override val topic: ActorTopic, override val service: ActorService) extends BasicActor {
+class RabbitMQPublisherActor(override val yellowPages: ActorRef, override val name: String,
+                             override val topic: ActorTopic, override val service: ActorService) extends BasicActor {
 
   override def receive: Receive = ({
     case PublishMessage(channel, exchange, routingKey, message) =>
