+++ src/main/scala/org/gammf/collabora/communication/actors/rabbitmq/RabbitMQUpdatesReceiverActor.scala
@@ -1,25 +1,22 @@
-package org.gammf.collabora.communication.actors
+package org.gammf.collabora.communication.actors.rabbitmq
 
 import akka.actor._
 import org.gammf.collabora.communication.Utils.CommunicationType
 import org.gammf.collabora.communication.messages._
 import org.gammf.collabora.util.UpdateMessage
-import org.gammf.collabora.yellowpages.ActorService.ActorService
-import org.gammf.collabora.yellowpages.actors.{BasicActor, YellowPagesActor}
-import play.api.libs.json.{JsError, JsSuccess, Json}
-import org.gammf.collabora.yellowpages.messages._
-import org.gammf.collabora.yellowpages.util.Topic._
-import org.gammf.collabora.yellowpages.util.Topic
+import org.gammf.collabora.yellowpages.ActorService.{ActorService, _}
 import org.gammf.collabora.yellowpages.TopicElement._
-import org.gammf.collabora.yellowpages.ActorService._
-
-import scala.concurrent.ExecutionContext.Implicits.global
+import org.gammf.collabora.yellowpages.actors.BasicActor
+import org.gammf.collabora.yellowpages.messages._
+import org.gammf.collabora.yellowpages.util.Topic
+import org.gammf.collabora.yellowpages.util.Topic._
+import play.api.libs.json.{JsError, JsSuccess, Json}
 
 /**
   * This is an actor that manages the reception of client updates.
   */
-class UpdatesReceiverActor(override val yellowPages: ActorRef, override val name: String,
-                           override val topic: ActorTopic, override val service: ActorService) extends BasicActor {
+class RabbitMQUpdatesReceiverActor(override val yellowPages: ActorRef, override val name: String,
+                                   override val topic: ActorTopic, override val service: ActorService) extends BasicActor {
 
   private[this] var subQueue: Option[String] = None
