+++ src/main/scala/org/gammf/collabora/database/actors/master/DBMasterMember.scala
@@ -3,7 +3,8 @@ package org.gammf.collabora.database.actors.master
 import akka.actor.ActorRef
 import akka.pattern.ask
 import org.gammf.collabora.communication.messages.{CommunicationMessage, PublishCollaborationInCollaborationExchange, PublishErrorMessageInCollaborationExchange, PublishNotificationMessage}
-import org.gammf.collabora.communication.actors.{NotificationsSenderActor, FirebaseActor}
+import org.gammf.collabora.communication.actors.FirebaseActor
+import org.gammf.collabora.communication.actors.rabbitmq.RabbitMQNotificationsSenderActor
 import org.gammf.collabora.database.messages._
 import org.gammf.collabora.util.{CollaborationMessage, ServerErrorCode, ServerErrorMessage, UpdateMessage, UpdateMessageTarget, UpdateMessageType}
 import org.gammf.collabora.yellowpages.ActorService.ActorService
@@ -91,7 +92,8 @@ class DBMasterMember(override val yellowPages: ActorRef, override val name: Stri
 }
 
 /**
-  * Message used internally by a [[DBMasterMember]] in order to send a message to a [[NotificationsSenderActor]].
+  * Message used internally by a [[DBMasterMember]] in order to send a message to a [[RabbitMQNotificationsSenderActor]].
+  *
   * @param insertMessage the message to be sent.
   */
 private case class SendInsertMemberNotificationMessage(insertMessage: InsertMemberMessage)
