+++ src/test/scala/org/gammf/collabora/communication/actors/RabbitMQCollaborationMembersActorTest.scala
@@ -13,7 +13,7 @@ import scala.concurrent.duration._
 import org.scalatest.concurrent.Eventually
 
 
-class CollaborationMembersActorTest extends TestKit (ActorSystem("CollaboraServer")) with WordSpecLike with Eventually with DefaultTimeout with Matchers with BeforeAndAfterAll with ImplicitSender {
+class RabbitMQCollaborationMembersActorTest extends TestKit (ActorSystem("CollaboraServer")) with WordSpecLike with Eventually with DefaultTimeout with Matchers with BeforeAndAfterAll with ImplicitSender {
 
   /*val factory = new ConnectionFactory()
   val connection:ActorRef = system.actorOf(ConnectionActor.props(factory), "rabbitmq")
