+++ src/test/scala/org/gammf/collabora/communication/actors/RabbitMQPublisherActorTest.scala
@@ -4,7 +4,7 @@ import akka.actor.{ ActorSystem}
 import akka.testkit.{DefaultTimeout, ImplicitSender, TestKit}
 import org.scalatest.{BeforeAndAfterAll, Matchers, WordSpecLike}
 
-class PublisherActorTest extends TestKit (ActorSystem("CollaboraServer")) with WordSpecLike with DefaultTimeout with Matchers with BeforeAndAfterAll with ImplicitSender {
+class RabbitMQPublisherActorTest extends TestKit (ActorSystem("CollaboraServer")) with WordSpecLike with DefaultTimeout with Matchers with BeforeAndAfterAll with ImplicitSender {
 
 
   override def beforeAll(): Unit ={
