+++ src/test/scala/org/gammf/collabora/communication/actors/RabbitMQSubscriberActorTest.scala
@@ -8,7 +8,7 @@ import org.gammf.collabora.communication.messages._
 import org.scalatest.{BeforeAndAfterAll, Matchers, WordSpecLike}
 
 
-class SubscriberActorTest extends TestKit (ActorSystem("CollaboraServer")) with WordSpecLike with DefaultTimeout with Matchers with BeforeAndAfterAll with ImplicitSender {
+class RabbitMQSubscriberActorTest extends TestKit (ActorSystem("CollaboraServer")) with WordSpecLike with DefaultTimeout with Matchers with BeforeAndAfterAll with ImplicitSender {
 
   /*private val EXCHANGE_NAME = "updates"
   private val ROUTING_KEY = ""
